Not Proud of Day Two.
I don't know enough about golang's data structures to be able to increment/decrement the position of scores to get the value of whatever I should play.

Like if there was an ordered k/v pair we could do `scores.indexOf("C") - scores[line[0]]` to get the value of what should be played for a loss.